refactor: Improve accounts API and centralize toolkit filtering
Addresses Coderabbit feedback by:

1.  Refactoring `hasProviderAccount` in `src/server/api/routers/accounts.ts`:
    *   Replaced `console.error` with TODOs for a proper logging service.
    *   Implemented a more robust provider mapping object for Clerk OAuth provider IDs.

2.  Centralizing Client-side Toolkit Filtering Logic:
    *   Created a new custom hook `src/app/_hooks/useFilteredToolkits.ts` to encapsulate the logic for determining `displayableToolkitIds` and `isClerkAccountsEnabled` based on the feature flag.
    *   Updated `NewWorkbenchForm`, `EditWorkbenchForm`, and the `ToolsSelect` component in `chat/input/tools.tsx` to use this hook. This reduces code duplication and ensures consistent toolkit availability display across the UI.
    *   Implicitly renames `useClerkAccounts` to `isClerkAccountsEnabled` in consuming components via the hook's return value, addressing a naming nitpick.
