+++ src/app/workbench/new/_components/new-workbench-form.tsx
@@ -22,10 +22,12 @@ import {
 } from "@/components/ui/dialog";
 import { Anvil, Plus } from "lucide-react";
 import { ToolkitIcons } from "@/components/toolkit/toolkit-icons";
-import { clientToolkits } from "@/toolkits/toolkits/client";
-import { useMemo } from "react"; // Added useMemo
-import { env } from "@/env"; // Added env
-import { Toolkits as ToolkitsEnum } from "@/toolkits/toolkits/shared"; // For enum access
+// import { clientToolkits } from "@/toolkits/toolkits/client"; // No longer directly needed if hook provides IDs
+// import { useMemo } from "react"; // Moved to hook
+// import { env } from "@/env"; // Moved to hook
+// import { Toolkits as ToolkitsEnum } from "@/toolkits/toolkits/shared"; // Moved to hook
+import { useFilteredToolkits } from "@/app/_hooks/useFilteredToolkits";
+
 
 export function NewWorkbenchForm() {
   const router = useRouter();
@@ -35,21 +37,8 @@ export function NewWorkbenchForm() {
     [],
   );
 
-  const useClerkAccounts = useMemo(() => env.NEXT_PUBLIC_FEATURE_EXTERNAL_ACCOUNTS_ENABLED, []);
-  const legacyToolkitsToHideWhenClerkActive: ToolkitsEnum[] = useMemo(() => [
-    ToolkitsEnum.Github,
-    ToolkitsEnum.GoogleCalendar,
-    ToolkitsEnum.Notion,
-    ToolkitsEnum.GoogleDrive,
-  ], []);
-
-  const displayableToolkitIds = useMemo(() => {
-    const allIds = Object.keys(clientToolkits) as ToolkitsEnum[];
-    if (useClerkAccounts) {
-      return allIds.filter(id => !legacyToolkitsToHideWhenClerkActive.includes(id));
-    }
-    return allIds;
-  }, [useClerkAccounts, legacyToolkitsToHideWhenClerkActive]);
+  const { displayableToolkitIds } = useFilteredToolkits();
+  // isClerkAccountsEnabled is also available from the hook if needed elsewhere
 
   const createMutation = api.workbenches.createWorkbench.useMutation({
     onSuccess: (workbench) => {
